Split models top-level file
